requirements.txt
Flask==3.1.*
gunicorn==23.0.*
rembg[cpu]==2.0.*
google-auth==2.35.*
requests==2.32.*
